The positions are now correct (for the moment, till we have better tests :-). Issue 12 and 15 still are not solved.
We try to store the numbers of whitespace between the tokens, so we can use this to create base_expr with the correct length. It is simpler to find these expressions with strpos(). The base_expr doesn't contain the part of the original statement, because some special characters are replaced with space characters. So we lose some tab or linefeeds and replace these characters with space characters. But the length is correct.

Tests have been corrected.

git-svn-id: https://php-sql-parser.googlecode.com/svn/trunk@83 44dcfd11-89b7-3d55-2063-5c365055f77f