Fix UnionProcessor instantiation
When processing a from portion of the query with a union join, the loading of the UnionProcessory requires (via the AbstractProcessor) that the constructor has the options object passed to it. This fixes that issue.

I did not conduct any unit tests, but doing this resolve a parsing error we had. Exact error was:

Argument 1 passed to PHPSQLParser\processors\AbstractProcessor::__construct() must be an instance of PHPSQLParser\Options, none given, called in /vendor/greenlion/php-sql-parser/src/PHPSQLParser/processors/FromProcessor.php on line 120 and defined (/vendor/greenlion/php-sql-parser/src/PHPSQLParser/processors/AbstractProcessor.php : 67);