fix issue #294: missing comma in cast
without this change, parsing and then re-creating a SQL query casting a
field as a decimal loses the comma in the decimal

for example `DECIMAL (16,2)` became `DECIMAL (16 2)`

added a `delim` in `ExpressionListProcessor`, when have a comma token,
as is done in `SelectProcessor#process`.

not sure if this is the right way to proceed,
and if it is it might need to be done in other places.