Added badly documented data type alias
Added a number of badly documented data type aliases which are valid data
types in MySQL and map to more familiar ones when a table is created
using them.

For example `MIDDLEINT` is an alias for `MEDIUMINT`.

This is needed if you're going to parse table create query digests from the
performance_schema as they sometimes reference the alias types.