Confusion between CHARACTER type and CHARACTER SET
In an attempt to fix the character set issue introduced by me in
https://github.com/greenlion/PHP-SQL-Parser/issues/355

This change separates out the CHARACTER check to do a look ahead
to see if the next keyword is SET and splits the logic appropriately.

An alternative would be to look back at the constructed expression
array ($expr) to see if a ExpressionType::DATA_TYPE has already been
defined.

This is an alternative fix to the fix proposed by
https://github.com/greenlion/PHP-SQL-Parser/pull/366