Added github actions to run phpunit tests (#378)
The ci.yml is configured to run the phpunit tests on any push to the
repository. This will run the tests for a matrix of PHP versions to
ensure all major versions supported are properly tests.

In addition, if the tests all pass, code coverage for the tests will be
generated and uploaded as an artifact of the run so they can be
downloaded and examined. To support this I needed to specify which
directories to be included in the code coverage in the phpunit.xml file.